refactor(Shell): rename to ts
