+++ components/shell/style.ts
