+++ components/shell/util.ts
